pyproject.toml
[tool.poetry]
name = "today_is_donuts_day_bot"
version = "0.1.0"
description = "It is reminding today is donuts day"
authors = ["Georgy Aleksahin"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.19"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
